+++ resharper/resharper-unity/src/Yaml/Psi/Resolve/UnityEventTargetReference.cs
@@ -1,11 +1,12 @@
-using System;
 using JetBrains.Annotations;
+using JetBrains.ReSharper.Plugins.Yaml.Psi.Parsing;
 using JetBrains.ReSharper.Plugins.Yaml.Psi.Tree;
 using JetBrains.ReSharper.Psi;
 using JetBrains.ReSharper.Psi.ExtensionsAPI;
 using JetBrains.ReSharper.Psi.ExtensionsAPI.Resolve;
 using JetBrains.ReSharper.Psi.Resolve;
 using JetBrains.ReSharper.Psi.Tree;
+using JetBrains.ReSharper.Resources.Shell;
 using JetBrains.Util;
 
 namespace JetBrains.ReSharper.Plugins.Unity.Yaml.Psi.Resolve
@@ -36,7 +37,14 @@ namespace JetBrains.ReSharper.Plugins.Unity.Yaml.Psi.Resolve
             if (!resolveResultWithInfo.Result.IsEmpty)
                 return resolveResultWithInfo;
 
-            return new ResolveResultWithInfo(EmptyResolveResult.Instance, ResolveErrorType.NOT_RESOLVED);
+            // TODO: Support references to scripts/event handlers in external packages
+            // Surprisingly, it's possible to have a reference to a script asset defined in a read-only package. We
+            // don't know anything about these assets, because read-only packages are not part of the C# project
+            // structure - they are compiled and added as assembly references. So we don't currently have a way to map
+            // an asset GUID back to a compiled class.
+            // See also UnityEventTargetReference
+//            return new ResolveResultWithInfo(EmptyResolveResult.Instance, ResolveErrorType.NOT_RESOLVED);
+            return new ResolveResultWithInfo(EmptyResolveResult.Instance, ResolveErrorType.IGNORABLE);
         }
 
         public override ISymbolTable GetReferenceSymbolTable(bool useReferenceName)
@@ -66,8 +74,16 @@ namespace JetBrains.ReSharper.Plugins.Unity.Yaml.Psi.Resolve
 
         public override IReference BindTo(IDeclaredElement element)
         {
-            // TODO: Handle rename
-            throw new NotImplementedException();
+            using (WriteLockCookie.Create(myOwner.IsPhysical()))
+            {
+                var text = YamlTokenType.NS_PLAIN_ONE_LINE.Create(element.ShortName);
+                if (myOwner.Text != null)
+                    LowLevelModificationUtil.ReplaceChildRange(myOwner.Text, myOwner.Text, text);
+                else
+                    LowLevelModificationUtil.AddChild(myOwner.Text, text);
+            }
+
+            return this;
         }
 
         private bool IsMonoBehaviourReference()