+++ rider/src/main/kotlin/com/jetbrains/rider/plugins/unity/notifications/AssetModeForceTextNotification.kt
@@ -31,8 +31,8 @@ class AssetModeForceTextNotification(private val propertiesComponent: Properties
 
         val message = """Some advanced integration features are unavailable when the Unity asset serialisation mode is not set to “Force Text”. Enable text serialisation to allow Rider to learn more about the structure of your scenes and assets.
             <ul style="margin-left:10px">
-              <a href="LearnMoreNavigateAction">Learn more</a>
-              <a href="doNotShow">Do not show</a> this notification for this solution.
+              <li><a href="LearnMoreNavigateAction">Learn more</a></li>
+              <li><a href="doNotShow">Do not show</a> this notification for this solution.</li>
             </ul>
             """
         val assetModeNotification = Notification(notificationGroupId.displayId, "Recommend switching to text asset serialisation mode", message, NotificationType.WARNING)